requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
aiohappyeyeballs==2.3.4
    # via aiohttp
aiohttp==3.10.1
    # via gcsfs
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
attrs==24.2.0
    # via aiohttp
bokeh==3.5.1
    # via dask
cachetools==5.4.0
    # via google-auth
certifi==2024.7.4
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   dask
    #   distributed
cloudpickle==3.0.0
    # via
    #   dask
    #   distributed
contourpy==1.2.1
    # via
    #   bokeh
    #   matplotlib
cycler==0.12.1
    # via matplotlib
dask[array,complete,dataframe,diagnostics,distributed]==2024.8.0
    # via
    #   ETL-Data-Pipeline (pyproject.toml)
    #   dask-expr
    #   distributed
dask-expr==1.1.10
    # via dask
decorator==5.1.1
    # via gcsfs
distributed==2024.8.0
    # via dask
fonttools==4.53.1
    # via matplotlib
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.6.1
    # via
    #   dask
    #   gcsfs
gcsfs==2024.6.1
    # via ETL-Data-Pipeline (pyproject.toml)
google-api-core==2.19.1
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.32.0
    # via
    #   gcsfs
    #   google-api-core
    #   google-auth-oauthlib
    #   google-cloud-core
    #   google-cloud-storage
google-auth-oauthlib==1.2.1
    # via gcsfs
google-cloud-core==2.4.1
    # via google-cloud-storage
google-cloud-storage==2.18.0
    # via
    #   ETL-Data-Pipeline (pyproject.toml)
    #   gcsfs
google-crc32c==1.5.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.1
    # via google-cloud-storage
googleapis-common-protos==1.63.2
    # via google-api-core
idna==3.7
    # via
    #   requests
    #   yarl
jinja2==3.1.4
    # via
    #   bokeh
    #   dask
    #   distributed
kiwisolver==1.4.5
    # via matplotlib
locket==1.0.0
    # via
    #   distributed
    #   partd
lz4==4.3.3
    # via dask
markupsafe==2.1.5
    # via jinja2
matplotlib==3.9.0
    # via ETL-Data-Pipeline (pyproject.toml)
msgpack==1.0.8
    # via distributed
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
numpy==2.0.1
    # via
    #   ETL-Data-Pipeline (pyproject.toml)
    #   bokeh
    #   contourpy
    #   dask
    #   matplotlib
    #   pandas
    #   pyarrow
    #   pyspark
oauthlib==3.2.2
    # via requests-oauthlib
packaging==24.1
    # via
    #   bokeh
    #   dask
    #   distributed
    #   matplotlib
pandas==2.2.2
    # via
    #   ETL-Data-Pipeline (pyproject.toml)
    #   bokeh
    #   dask
    #   dask-expr
    #   pyspark
partd==1.4.2
    # via dask
pillow==10.4.0
    # via
    #   bokeh
    #   matplotlib
polars==1.5.0
    # via ETL-Data-Pipeline (pyproject.toml)
proto-plus==1.24.0
    # via google-api-core
protobuf==5.27.3
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
psutil==6.0.0
    # via distributed
py4j==0.10.9.7
    # via pyspark
pyarrow==17.0.0
    # via
    #   ETL-Data-Pipeline (pyproject.toml)
    #   dask
    #   dask-expr
    #   pyspark
pyarrow-hotfix==0.6
    # via dask
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
pydantic==2.8.2
    # via ETL-Data-Pipeline (pyproject.toml)
pydantic-core==2.20.1
    # via pydantic
pyparsing==3.1.2
    # via matplotlib
pyspark[sql]==3.5.1
    # via ETL-Data-Pipeline (pyproject.toml)
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2024.1
    # via pandas
pyyaml==6.0.2
    # via
    #   bokeh
    #   dask
    #   distributed
requests==2.32.3
    # via
    #   gcsfs
    #   google-api-core
    #   google-cloud-storage
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
six==1.16.0
    # via python-dateutil
sortedcontainers==2.4.0
    # via distributed
tblib==3.0.0
    # via distributed
toolz==0.12.1
    # via
    #   dask
    #   distributed
    #   partd
tornado==6.4.1
    # via
    #   bokeh
    #   distributed
typing-extensions==4.12.2
    # via
    #   pydantic
    #   pydantic-core
tzdata==2024.1
    # via pandas
urllib3==2.2.2
    # via
    #   distributed
    #   requests
xyzservices==2024.6.0
    # via bokeh
yarl==1.9.4
    # via aiohttp
zict==3.0.0
    # via distributed
